requirements.txt
streamlit
pandas
nltk
joblib
python-dotenv
openai
sqlite3
```